test: factor out common_functions.py
